re add removed file
