+++ src/main/java/ch/heigvd/amt/mvcProject/ui/web/user/LoginRenderer.java
@@ -1,0 +1,16 @@
+package ch.heigvd.amt.mvcProject.ui.web.user;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@WebServlet(name = "LoginRenderer", urlPatterns = "/login")
+public class LoginRenderer extends HttpServlet {
+
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        request.getRequestDispatcher("/WEB-INF/views/login.jsp").forward(request, response);
+    }
+}
