+++ src/main/java/ch/heigvd/amt/mvcProject/ui/web/user/RegisterRenderer.java
@@ -1,0 +1,15 @@
+package ch.heigvd.amt.mvcProject.ui.web.user;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@WebServlet(name = "RegisterRenderer", urlPatterns = "/register")
+public class RegisterRenderer extends HttpServlet {
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        request.getRequestDispatcher("/WEB-INF/views/register.jsp").forward(request, response);
+    }
+}
