+++ src/main/java/ch/heigvd/amt/mvcProject/ui/web/user/handler/LoginRequestHandler.java
@@ -1,5 +1,4 @@
-package ch.heigvd.amt.mvcProject.ui.web.user;
-
+package ch.heigvd.amt.mvcProject.ui.web.user.handler;
 
 import ch.heigvd.amt.mvcProject.application.ServiceRegistry;
 import ch.heigvd.amt.mvcProject.application.user.UserCommand;
@@ -12,8 +11,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "RegisterRequestHandler", urlPatterns = "/request.register")
-public class RegisterRequestHandler extends HttpServlet {
+@WebServlet(name = "LoginRequestHandler", urlPatterns = "/request.login")
+public class LoginRequestHandler extends HttpServlet {
 
     private ServiceRegistry serviceRegistry;
     private UserFacade userFacade;
@@ -27,27 +26,17 @@ public class RegisterRequestHandler extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        UserCommand command = UserCommand.builder()
+                .username(req.getParameter("username"))
+                .email("")
+                .password(req.getParameter("password"))
+                .build();
 
-
-        String password = req.getParameter("password");
-        String cPasssword = req.getParameter("cPassword");
-
-        if (password.equals(cPasssword)) {
-
-            UserCommand command = UserCommand.builder()
-                    .username(req.getParameter("username"))
-                    .email(req.getParameter("email"))
-                    .password(password)
-                    .build();
-
-
-            userFacade.addNewUser(command);
-
+        if (userFacade.isUserExist(command.getUserId())) {
             resp.sendRedirect(getServletContext().getContextPath());
-
         } else {
-            resp.sendRedirect(getServletContext()
-                    .getContextPath() + "/register?error='Your password and your confirmation aren't the same'");
+            resp.sendRedirect(
+                    getServletContext().getContextPath() + "/login?error=Your e-mail or your password is incorrect");
         }
     }
 }
